third_party/WebKit/LayoutTests/scrollingcoordinator/plugin-with-wheel-handler-expected.txt

This test ensures that a plugin which wants to receive wheel events is included in the non-fast scrollable region.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.nonFastScrollableRects(document).length is 1
PASS successfullyParsed is true

TEST COMPLETE
